+++ crates/test_util/src/lib.rs
@@ -26,8 +26,41 @@ pub use self::typing::{
     DecodeExampleData, DurationExampleData, ExampleData, IntonationExampleData,
 };
 
-pub const ONNXRUNTIME_DYLIB_PATH: &str =
-    include_str!(concat!(env!("OUT_DIR"), "/onnxruntime-dylib-path.txt"));
+pub const ONNXRUNTIME_DYLIB_PATH: &str = {
+    #[cfg(any(target_os = "linux", target_os = "macos"))]
+    macro_rules! version {
+        () => {
+            include_str!("../../../onnxruntime-version.txt")
+        };
+    }
+
+    #[cfg(target_os = "windows")]
+    macro_rules! lib_versioned_filename {
+        () => {
+            "onnxruntime.dll"
+        };
+    }
+
+    #[cfg(target_os = "linux")]
+    macro_rules! lib_versioned_filename {
+        () => {
+            concat!("libonnxruntime.so.", version!())
+        };
+    }
+
+    #[cfg(target_os = "macos")]
+    macro_rules! lib_versioned_filename {
+        () => {
+            concat!("libonnxruntime.", version!(), ".dylib")
+        };
+    }
+
+    concat!(
+        env!("CARGO_MANIFEST_DIR"),
+        "/../../target/voicevox_core/downloads/onnxruntime/",
+        lib_versioned_filename!(),
+    )
+};
 
 pub const OPEN_JTALK_DIC_DIR: &str = concat!(
     env!("CARGO_MANIFEST_DIR"),