+++ crates/voicevox_core_c_api/tests/e2e/main.rs
@@ -22,7 +22,10 @@ fn main() -> anyhow::Result<()> {
     enum TestContext {}
 
     impl assert_cdylib::TestContext for TestContext {
-        const FEATURES: &'static [&'static str] = &["load-onnxruntime"];
+        const FEATURES: &'static [&'static str] = &[
+            "voicevox_core/buildtime-download-onnxruntime",
+            "load-onnxruntime",
+        ];
         const TARGET_DIR: &'static str = "../../target";
         const CDYLIB_NAME: &'static str = "voicevox_core";
         const RUNTIME_ENVS: &'static [(&'static str, &'static str)] =